# public/swagger/api.yml
openapi: "3.0.3"
info:
  title: transmitter API
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local server
  - url: https://liscitransmitter.live
    description: Production server
paths:
  /api/checkKey:
    get:
      summary: Check API Key
      description: Validates the provided API key.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
      responses:
        "200":
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
                example: Valid username
        "401":
          description: Invalid API key
  /api/getPosts:
    get:
      summary: Get Posts
      description: Retrieves a list of posts.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
        - in: query
          name: page
          schema:
            type: integer
          description: The page number to retrieve.
        - in: query
          name: filter
          schema:
            type: string
          description: The filter to apply. Can be a JSON string representing the filter object.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    user:
                      type: string
                    title:
                      type: string
                    content:
                      type: string
                    mediaPath:
                      type: string
                    type:
                      type: string
                    likes:
                      type: integer
        "401":
          description: Invalid API key
  /api/getMostRecentPost:
    get:
      summary: Get Most Recent Post
      description: Retrieves the most recent post.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
        - in: query
          name: filter
          schema:
            type: string
          description: The filter to apply. Can be a JSON string representing the filter object.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  user:
                    type: string
                  title:
                    type: string
                  sections:
                    type: array
                    items:
                      type: string
                  mediaPath:
                    type: string
                  type:
                    type: string
                  likes:
                    type: integer
                  liked:
                    type: boolean
        "401":
          description: Invalid API key
        "404":
          description: No posts found
        "400":
          description: Invalid filter object
  /api/getPostPages:
    get:
      summary: Get Post Pages
      description: Retrieves the number of pages of posts.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
        - in: query
          name: filter
          schema:
            type: string
            enum:
              - all
              - posts
              - news
          description: The filter to apply.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: integer
                    description: The number of pages.
        "401":
          description: Invalid API key
  /api/getComments:
    get:
      summary: Get Comments
      description: Retrieves comments for a specific post.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
        - in: query
          name: postID
          schema:
            type: string
          description: The ID of the post to retrieve comments for.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    user:
                      type: string
                    content:
                      type: string
        "401":
          description: Invalid API key
  /api/createPost:
    post:
      summary: Create Post
      description: Creates a new post.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                type:
                  type: string
                permissions:
                  type: string
                mediaPath:
                  type: string
      responses:
        "200":
          description: Successful response
        "401":
          description: Invalid API key
        "403":
          description: Forbidden
        "500":
          description: Internal server error
  /api/createComment:
    post:
      summary: Create Comment
      description: Creates a new comment on a post.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                postID:
                  type: string
                content:
                  type: string
                permissions:
                  type: string
      responses:
        "200":
          description: Successful response
        "401":
          description: Invalid API key
        "500":
          description: Internal server error
  /api/likePost:
    post:
      summary: Like Post
      description: Likes a post.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                postID:
                  type: string
      responses:
        "200":
          description: Successful response
        "401":
          description: Invalid API key
        "500":
          description: Internal server error
  /api/sendPush:
    post:
      summary: Send Push Notification
      description: Sends a push notification.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userID:
                  type: string
                title:
                  type: string
                body:
                  type: string
                icon:
                  type: string
                badge:
                  type: string
                urgency:
                  type: string
      responses:
        "200":
          description: Successful response
        "401":
          description: Invalid API key
        "403":
          description: Forbidden
        "500":
          description: Internal server error
  /api/createCitation:
    post:
      summary: Create Citation
      description: Creates a new citation.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                author:
                  type: string
                content:
                  type: string
      responses:
        "200":
          description: Successful response
        "401":
          description: Invalid API key
        "500":
          description: Internal server error
  /api/getCitations:
    post:
      summary: Get Citations
      description: Retrieves a list of citations.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                number:
                  type: integer
                filter:
                  type: string
                  description: The filter to apply. Can be a JSON string representing the filter object.
                sort:
                  type: string
                  description: The sort order. Can be a JSON string representing the sort object.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    author:
                      type: string
                    content:
                      type: string
        "401":
          description: Invalid API key
        "400":
          description: Invalid filter or sort object
        "500":
          description: Internal server error
  /api/getMostRecentCitation:
    get:
      summary: Get Most Recent Citation
      description: Retrieves the most recent citation.
      parameters:
        - in: header
          name: x-api-key
          required: true
          schema:
            type: string
          description: The API key to validate.
        - in: query
          name: filter
          schema:
            type: string
          description: The filter to apply. Can be a JSON string representing the filter object.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  author:
                    type: string
                  content:
                    type: string
        "401":
          description: Invalid API key
        "400":
          description: Invalid filter object
        "500":
          description: Internal server error